+++ initialize/default_files.go
@@ -213,7 +213,7 @@ Usually this group will have the most number of keys of all of the groups.`
 		accountTypeDefinition.DefaultNumber = 25
 		accountTypeDefinition.DefaultTokens = 9999999999
 		accountTypeDefinition.DefaultBond = 0
-		accountTypeDefinition.Perms = map[string]int{
+		accountTypeDefinition.Perms = map[string]int64{
 			"root":            0,
 			"send":            1,
 			"call":            1,
@@ -243,7 +243,7 @@ within this group, although that design is up to you.`
 		accountTypeDefinition.DefaultNumber = 6
 		accountTypeDefinition.DefaultTokens = 9999999999
 		accountTypeDefinition.DefaultBond = 0
-		accountTypeDefinition.Perms = map[string]int{
+		accountTypeDefinition.Perms = map[string]int64{
 			"root":            0,
 			"send":            1,
 			"call":            1,
@@ -275,7 +275,7 @@ in the system.`
 		accountTypeDefinition.DefaultNumber = 7
 		accountTypeDefinition.DefaultTokens = 9999999999
 		accountTypeDefinition.DefaultBond = 9999999998
-		accountTypeDefinition.Perms = map[string]int{
+		accountTypeDefinition.Perms = map[string]int64{
 			"root":            0,
 			"send":            0,
 			"call":            0,
@@ -306,7 +306,7 @@ If you are making a more complex chain, don't use this account type.`
 		accountTypeDefinition.DefaultNumber = 1
 		accountTypeDefinition.DefaultTokens = 99999999999999
 		accountTypeDefinition.DefaultBond = 9999999999
-		accountTypeDefinition.Perms = map[string]int{
+		accountTypeDefinition.Perms = map[string]int64{
 			"root":            1,
 			"send":            1,
 			"call":            1,
@@ -338,7 +338,7 @@ similar to a network administrator in other data management situations.`
 		accountTypeDefinition.DefaultNumber = 3
 		accountTypeDefinition.DefaultTokens = 9999999999
 		accountTypeDefinition.DefaultBond = 0
-		accountTypeDefinition.Perms = map[string]int{
+		accountTypeDefinition.Perms = map[string]int64{
 			"root":            1,
 			"send":            1,
 			"call":            1,
@@ -370,7 +370,7 @@ func defaultChainTypes(chainType string) *definitions.ChainType {
 
 		chainTypeDefinition.Name = "simplechain"
 		chainTypeDefinition.Description = `A simple chain type will build a single node chain. This chain type is usefulfor quick and easy prototyping. It should not be used for anything more than the most simple prototyping as it only has one node and the keys to that node could get lost or compromised and the chain would thereafter become useless.`
-		chainTypeDefinition.AccountTypes = map[string]int{
+		chainTypeDefinition.AccountTypes = map[string]int64{
 			"Full":        1,
 			"Developer":   0,
 			"Participant": 0,
@@ -382,7 +382,7 @@ func defaultChainTypes(chainType string) *definitions.ChainType {
 
 		chainTypeDefinition.Name = "sprawlchain"
 		chainTypeDefinition.Description = `A sprawlchain type has a little bit of everything. Modify as necessary for your ecosystem application. Will tolerate three nodes down. As with other chains, Validator accounts ought to go on the cloud. No Full accounts are provided since these are prefered for quick development only.`
-		chainTypeDefinition.AccountTypes = map[string]int{
+		chainTypeDefinition.AccountTypes = map[string]int64{
 			"Full":        0,
 			"Developer":   10,
 			"Participant": 20,
@@ -394,7 +394,7 @@ func defaultChainTypes(chainType string) *definitions.ChainType {
 
 		chainTypeDefinition.Name = "adminchain"
 		chainTypeDefinition.Description = `An adminchain type has settings for prototyping a larger chain from a sysadmin point of view. With four Validator and three Full account_types, at minimum of five nodes must be up for consensus to happen. This account combination is what we use to test long running chains on our CI system.`
-		chainTypeDefinition.AccountTypes = map[string]int{
+		chainTypeDefinition.AccountTypes = map[string]int64{
 			"Full":        3,
 			"Developer":   1,
 			"Participant": 1,
@@ -406,7 +406,7 @@ func defaultChainTypes(chainType string) *definitions.ChainType {
 
 		chainTypeDefinition.Name = "demochain"
 		chainTypeDefinition.Description = `A demo chain is useful for setting up proof of concept demonstration chains. It is a quick and easy way to have multi-validator, multi-developer, multi-participant chains set up for your application. This chain will tolerate 2 validators becoming byzantine or going off-line while still moving forward. You should utilize 7 different cloud instances and deploy one of the validator chain directories to each.`
-		chainTypeDefinition.AccountTypes = map[string]int{
+		chainTypeDefinition.AccountTypes = map[string]int64{
 			"Full":        0,
 			"Developer":   5,
 			"Participant": 20,
@@ -418,7 +418,7 @@ func defaultChainTypes(chainType string) *definitions.ChainType {
 
 		chainTypeDefinition.Name = "gochain"
 		chainTypeDefinition.Description = `A gochain type will build a three node chain. It is a quick and easy way to get started with a multi-validator chain. The Full account_type includes validator and deploy permissions, allowing for experimentation with setting up a chain and halting it by taking down a single node. This Full account should be deployed on your local machine and cloud nodes should have only Validator accounts. Use for prototyping only.`
-		chainTypeDefinition.AccountTypes = map[string]int{
+		chainTypeDefinition.AccountTypes = map[string]int64{
 			"Full":        1,
 			"Developer":   0,
 			"Participant": 0,
